Encapsulated the rightStackPanel's (now referred to as Portfolio List) operations.
Issue #28
