+++ CryptFolio-UWP/CryptFolio-UWP.csproj
@@ -101,6 +101,7 @@
     <Compile Include="MainPage.xaml.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
+    <Compile Include="PortfolioList.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Splash.xaml.cs">
       <DependentUpon>Splash.xaml</DependentUpon>
