+++ CryptFolio-UWP/MainPage.xaml.cs
@@ -1,12 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Windows.UI;
-using Windows.UI.Text;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Media;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=402352&clcid=0x409
 
@@ -19,15 +16,13 @@ namespace CryptFolio
     {
         private Dictionary<string, string> nameDictionary;
         private string selectedCurrency;
-        private double addedAmount, totalInvestmentValue;
-        private TickerJSONResult result;
+        private PortfolioList portfolioList;
 
         public MainPage()
         {
             this.InitializeComponent();
             this.nameDictionary = App.apiObj.GetNameDictionary();
-
-
+            this.portfolioList = new PortfolioList();
         }
         private void AddCurrenciesToMarketView()
         {
@@ -53,10 +48,10 @@ namespace CryptFolio
 
                 // get result from list if it exists
                 var timeSinceLastUpdated = API_RECOMMENDED_LIMIT + 1; // initialize to higher than API limit
-                result = App.apiObj.RetrieveJSONById(id);
+                portfolioList.result = App.apiObj.RetrieveJSONById(id);
 
-                if (App.jsonList != null && result != null)
-                    timeSinceLastUpdated = currentTime - Convert.ToInt64(result.last_updated);
+                if (App.jsonList != null && portfolioList.result != null)
+                    timeSinceLastUpdated = currentTime - Convert.ToInt64(portfolioList.result.last_updated);
 
                 if (timeSinceLastUpdated > API_RECOMMENDED_LIMIT)
                 {
@@ -65,7 +60,8 @@ namespace CryptFolio
                 }
                 else
                 {
-                    DisplayCurrencyStats(id, selectedItem);
+                    portfolioList.DisplayCurrencyStats(id, selectedItem, ref stackPanelRight);
+                    UpdateTotalValue();
                 }
             }
             catch (Exception ex)
@@ -81,14 +77,15 @@ namespace CryptFolio
             {
                var u = Dispatcher.RunAsync(Windows.UI.Core.CoreDispatcherPriority.Normal, () =>
                {
-                   result = a.Result;
+                   portfolioList.result = a.Result;
 
-                   if (result != null)
+                   if (a.Result != null)
                    {
                        // update last_updated value of the currency
                        var index = App.jsonList.FindIndex(x => x.id == id);
                        App.jsonList[index].last_updated = currentTime;
-                       DisplayCurrencyStats(id, selectedItem);
+                       portfolioList.DisplayCurrencyStats(id, selectedItem, ref stackPanelRight);
+                       UpdateTotalValue();
                    }
                    else
                        DisplayAPIError();
@@ -109,129 +106,6 @@ namespace CryptFolio
             stackPanelForError.Children.Add(errorTextBlock);
 
             stackPanelRight.Children.Add(stackPanelForError);
-        }
-
-        private void DisplayCurrencyStats(string ticker, string displayName)
-        {
-            TextBlock currencyLabel, usdLabel, btcLabel, userAmountLabel, userUSDValueLabel;
-            CustomStackPanel stackPanelToUpdate;
-
-            string
-                stackPanelStr = ticker + "StackPanel",
-                labelStr = ticker + "Label",
-                labelContentStr = displayName + " Stats:",
-                labelUSDStr = ticker + "USD",
-                labelUSDContentStr = "USD Price: $" + result.price_usd,
-                labelBTCStr = ticker + "BTC",
-                labelBTCContentStr = "BTC Price: " + result.price_btc,
-                labelUserAmountStr = ticker + "UserAmount",
-                labelUserAmountContentStr = "Amount you own: ",
-                labelUserUSDValueStr = ticker + "UserUSDVal",
-                labelUserUSDValueContentStr = "Your estimated USD value: ";
-
-            // add label if currency not already displayed
-            stackPanelToUpdate = GetChildOfStackPanel(stackPanelRight, stackPanelStr) as CustomStackPanel;
-
-            if (stackPanelToUpdate == null)
-            {
-                // create new StackPanel to hold labels
-                stackPanelToUpdate = new CustomStackPanel
-                {
-                    Name = stackPanelStr,
-                    Margin = new Thickness(5, 0, 5, 5)
-                };
-
-                // create and add currency label
-                currencyLabel = new TextBlock
-                {
-                    FontWeight = FontWeights.Bold,
-                    FontSize = 32,
-                    Text = labelContentStr,
-                    Name = labelStr
-                };
-                stackPanelToUpdate.Children.Add(currencyLabel);
-
-                // create and add usdLabel
-                usdLabel = new TextBlock
-                {
-                    Name = labelUSDStr,
-                    FontSize = 22,
-                    Text = labelUSDContentStr
-                };
-                stackPanelToUpdate.Children.Add(usdLabel);
-
-                // create and add btcLabel
-                btcLabel = new TextBlock
-                {
-                    Name = labelBTCStr,
-                    FontSize = 22,
-                    Text = labelBTCContentStr
-                };
-                stackPanelToUpdate.Children.Add(btcLabel);
-
-                // create and add user holding amount label
-                userAmountLabel = new TextBlock
-                {
-                    Name = labelUserAmountStr,
-                    FontSize = 22,
-                    Text = labelUserAmountContentStr + addedAmount.ToString()
-                };
-                stackPanelToUpdate.Children.Add(userAmountLabel);
-
-                // create and add userUSDValue label
-                userUSDValueLabel = new TextBlock
-                {
-                    Name = labelUserUSDValueStr,
-                    FontSize = 22,
-                    Text = UpdateUserUSDValueContentStr(ref stackPanelToUpdate, this.result.price_usd, labelUserUSDValueContentStr),
-                    Foreground = new SolidColorBrush(Colors.ForestGreen)
-                };
-                stackPanelToUpdate.Children.Add(userUSDValueLabel);
-
-                stackPanelRight.Children.Add(stackPanelToUpdate);
-            }
-            else
-            {
-                // update user value
-                stackPanelToUpdate.UserValue = addedAmount * Convert.ToDouble(this.result.price_usd);
-
-                // update price labels
-                usdLabel = GetChildOfStackPanel(stackPanelToUpdate, labelUSDStr) as TextBlock;
-                usdLabel.Text = labelUSDContentStr;
-                btcLabel = GetChildOfStackPanel(stackPanelToUpdate, labelBTCStr) as TextBlock;
-                btcLabel.Text = labelBTCContentStr;
-                userUSDValueLabel = GetChildOfStackPanel(stackPanelToUpdate, labelUserUSDValueStr) as TextBlock;
-                userUSDValueLabel.Text = UpdateUserUSDValueContentStr(ref stackPanelToUpdate, this.result.price_usd, labelUserUSDValueContentStr);
-                userAmountLabel = GetChildOfStackPanel(stackPanelToUpdate, labelUserAmountStr) as TextBlock;
-                userAmountLabel.Text = labelUserAmountContentStr + addedAmount.ToString();
-            }
-
-            // update total values
-            UpdateTotalValue();
-        }
-
-        private void UpdateTotalValue()
-        {
-            totalInvestmentValue = 0.0;
-            foreach (CustomStackPanel child in stackPanelRight.Children)
-            {
-                totalInvestmentValue += child.UserValue;
-            }
-
-            textBlockTotalValue.Text = "Total investment value: $" + totalInvestmentValue.ToString();
-        }
-
-        private FrameworkElement GetChildOfStackPanel(StackPanel parentPanel, string name)
-        {
-            foreach (FrameworkElement child in parentPanel.Children)
-            {
-                if (child.Name == name)
-                {
-                    return child;
-                }
-            }
-
-            return null;
         }
 
         private void AutoSuggestBox_TextChanged(AutoSuggestBox sender, AutoSuggestBoxTextChangedEventArgs args)
@@ -299,12 +173,12 @@ namespace CryptFolio
             {
                 try
                 {
-                    addedAmount = Convert.ToDouble(amountOwnedBox.Text);
-                    addedAmount = Math.Round(addedAmount, 6);
+                    portfolioList.addedAmount = Convert.ToDouble(amountOwnedBox.Text);
+                    portfolioList.addedAmount = Math.Round(portfolioList.addedAmount, 6);
                 }
                 catch (Exception)
                 {
-                    amountOwnedBox.Text = addedAmount.ToString();
+                    amountOwnedBox.Text = portfolioList.addedAmount.ToString();
                     FlyoutBase.ShowAttachedFlyout(amountOwnedBox);
                 }
             }
@@ -322,21 +196,11 @@ namespace CryptFolio
             AddCurrenciesToMarketView();
         }
 
-        private string UpdateUserUSDValueContentStr(ref CustomStackPanel sP, string priceStr, string contentStr)
+        private void UpdateTotalValue()
         {
-            // update user value
-            sP.UserValue = addedAmount * Convert.ToDouble(priceStr);
-
-            if (addedAmount == 0.0)
-            {
-                contentStr += "N/A";
-            }
-            else
-            {
-                contentStr += "$" + sP.UserValue.ToString();
-            }
-
-            return contentStr;
+            textBlockTotalValue.Text = "Total investment value: $" + portfolioList.CalculateTotalInvestmentValue(ref stackPanelRight).ToString();
         }
     }
+
+    
 }
